Move issues components into sub-directory and move major components to their own files.
