+++ src/Components/SideDrawerPanels.jsx
@@ -4,8 +4,9 @@ import Typography from '@mui/material/Typography'
 import {makeStyles, useTheme} from '@mui/styles'
 import useStore from '../store/useStore'
 import ItemProperties from './ItemProperties'
-import {IssuesNavBar, Issues} from './IssuesControl'
 import {TooltipIconButton} from './Buttons'
+import Issues from './issues/Issues'
+import IssuesNavBar from './issues/IssuesNavBar'
 import CloseIcon from '../assets/2D_Icons/Close.svg'
